+++ Sources/Readability/Internal/SiteRules/DeclarativeSiteRule.swift
@@ -1,3 +1,6 @@
+// Copyright 2026 Binary Dreams, LLC.
+// SPDX-License-Identifier: Apache-2.0
+
 import Foundation
 import SwiftSoup